src/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cmdstanpy==1.2.4
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
dill==0.3.8
distlib==0.3.8
exceptiongroup==1.2.2
fastapi==0.115.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.9.0
h11==0.14.0
holidays==0.56
identify==2.6.1
idna==3.8
importlib_resources==6.4.4
isort==5.13.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mccabe==0.7.0
mpmath==1.3.0
networkx==3.2.1
nodeenv==1.9.1
numpy==1.24.3
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pillow==10.4.0
platformdirs==4.3.3
plotly==5.24.0
pre-commit==3.8.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prophet==1.1.5
pydantic==2.9.2
pydantic_core==2.23.4
pylint==3.2.7
pyparsing==3.1.4
pystan==2.19.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==72.1.0
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
stanio==0.5.1
starlette==0.38.5
statsmodels==0.14.2
sympy==1.13.2
tenacity==9.0.0
threadpoolctl==3.5.0
tomli==2.0.1
tomlkit==0.13.2
torch==2.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
virtualenv==20.26.4
xgboost==2.1.1
yfinance==0.2.43
zipp==3.20.1
